Rename `replicon_tick` module into `tick`
