+++ src/server/server_tick.rs
@@ -1,7 +1,7 @@
 use bevy::prelude::*;
 use serde::{Deserialize, Serialize};
 
-use crate::core::replicon_tick::RepliconTick;
+use crate::core::tick::RepliconTick;
 
 /// Stores current [`RepliconTick`].
 ///
